+++ odoo/local-src/sf_rma/models/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
+from . import account_invoice
 from . import models
 from . import mrp_repair
 from . import picking