+++ HISTORY.rst
@@ -23,6 +23,7 @@ latest (unreleased)
 
 * BSSFL-488: Change RMA rules for Picking generation. The conditions for
   generating the procurement order have been changed if he is linking to a rma.
+* BSSFL-499: Generate picking (on conditions) when down payment is paid
 
 **Bugfixes**